发布0.1beta版本

+++ src/easygqcnn/grasp_mapper.py
@@ -1,6 +1,7 @@
 import logging
 import numpy as np
 import cv2
+import scipy.misc as sm
 
 
 class GraspMapper(object):
@@ -41,6 +42,7 @@ class GraspMapper(object):
             image_T = self.transform(image, g.center_float, g.angle)
             # 缩放到目标大小
             image_T = self.resize_image(image_T, w / g.width_px)
+            # image_T = self.resize_image(image_T, 1./3.0)
             # 裁剪到最终大小
             image_T = self.crop_image(image_T, out_size)
             out_image[i] = image_T
@@ -89,7 +91,8 @@ class GraspMapper(object):
     @staticmethod
     def resize_image(image, resize_rate):
         """ 缩放图片大小 """
-        image_size = np.array(image.shape[:2][::-1]).astype(np.int)
+        image_size = np.array(image.shape[:2][::-1])
         out_size = np.ceil(image_size * resize_rate).astype(np.int)
         image_out = cv2.resize(image, tuple(out_size))
+        # image_out = sm.imresize(image, resize_rate, interp='bilinear', mode='F')
         return image_out
